Alguns ajustes de texturas botoes agora com textura de solecionado/não selecionado, mostra vida e pontuação na tela, menu jogar mostrando 1 ou 2 bonecos

+++ sources/Menu/MenuSalvarPontuacao.cpp
@@ -99,7 +99,7 @@ namespace Estados
             std::ofstream arquivo("../resources/leaderboard.txt", std::ios::app);
             if (arquivo.is_open())
             {
-                arquivo << nomeString << " " << Entidades::Personagem::Jogador::getPontuacao() << std::endl;
+                arquivo << nomeString << " " << (int)Entidades::Personagem::Jogador::getPontuacao() << std::endl;
                 arquivo.close();
             }
         }
